perl_rules: Fix runfiles lookup in sandbox
With the command line flag --nobuild_runfile_links, the tool wrapper
for perl cannot find the runfiles anymore. It searches outside the
sandbox due to the resolution of the symlinks. But with the flag
applied, the runfiles tree is not created outside the sandbox.

With this change, the wrapper scripts also works if the runfiles tree
is omitted.