WIP: Introduce the idea of toolchains and add `perl_xs` rule
This is a WIP that introduces the idea of toolchains to the rules, and
uses this to download a toolchain for perl.

A downloadable toolchain is provided via the `deps.bzl` file to allow
for a hermetic in tree toolchain of perl. This toolchain is a
redistributable perl from `skaji` although it could probably be
reproduced under a different banner for practical reasons of toolchain
compliance.

Using the download toolchain is as follows:

```
load(
    "@io_bazel_rules_perl//perl:deps.bzl",
    "perl_register_toolchains",
    "perl_rules_dependencies",
)

perl_rules_dependencies()
perl_register_toolchains()
```

As part of these changes the perl launcher script was replaced with a
simpler shell script that covers launching `perl_binary` and `perl_test`
targets which is aware of the given bazel toolchain.

Along with these changes the `perl_xs` rule is introduced that allows
for perl extension modules to be _crudely_ compiled. This was needed by
us at $DAYJOB to get a bunch of cpan modules to be built that use
various C extensions (the test of these changes was `IO-Tty`).