Introduce the idea of toolchains and add `perl_xs` rule (#13)
This is the beginning of the idea of toolchains to the rules, and
uses this to download a toolchain for perl.

A downloadable toolchain is provided via the `deps.bzl` file to allow
for a hermetic in tree toolchain of perl. This toolchain is a
redistributable perl from `skaji` although it could probably be
reproduced under a different banner for practical reasons of toolchain
compliance.

Using the download toolchain is as follows:

```
load(
    "@io_bazel_rules_perl//perl:deps.bzl",
    "perl_register_toolchains",
    "perl_rules_dependencies",
)

perl_rules_dependencies()
perl_register_toolchains()
```

As part of these changes the perl launcher script was replaced with a
simpler shell script that covers launching `perl_binary` and `perl_test`
targets which is aware of the given bazel toolchain.

Along with these changes the `perl_xs` rule is introduced that allows
for perl extension modules to be _crudely_ compiled. This was needed by
us at $DAYJOB to get a bunch of cpan modules to be built that use
various C extensions (the test of these changes was `IO-Tty`).

This change attempts to be backwards compatible so that all current
tests pass.

New examples have been added to show the inclusion of a Pure Perl 
and complied module examples/tests.

Mac compatibility should be preserved and extended to be able
to include Pure Perl modules, however, modules that require compilation
are not supported at this time.

https://github.com/GregBowyer/ - He started it all ... damn him!

https://github.com/JimmyTheJudd/ - JimmTheJudd was a huge
help in getting the rules this far.  His commit got squished,
but he defintely deserves the shout-out.

Co-authored-by: Erik Tank <lefthandskeletonkey@gmail.com>